requirements.txt
absl-py==1.4.0
numpy==1.25.1
protobuf==3.20.2
dm-env==1.6
python-dotenv==1.0.0
google-auth==1.35.0
grpcio==1.50.0
dm-acme @ git+https://github.com/deepmind/acme@fb87fd53ce7250584dc94c481cba949b857ff255
dm-launchpad==0.5.2
chex==0.1.82
dm-haiku==0.0.9
tensorflow==2.8.0
dm-reverb==0.7.2
gym==0.25.0
matplotlib==3.7.0
tensorflow-probability==0.15.0
dm-sonnet==2.0.1
rlax==0.1.6
optax==0.1.5
fastavro==1.7.0
google-cloud-bigquery==2.34.4
google-cloud-storage==1.44.0
dsub==0.4.7